chore: rename community repository file
