+++ lib/src/features/community/community.dart
@@ -1,7 +1,7 @@
 export 'domain/enums/community_action.dart';
 export 'domain/models/community_details.dart';
 export 'package:thunder/src/foundation/primitives/models/thunder_community.dart';
-export 'data/repositories/community_repository_impl.dart';
+export 'data/repositories/community_repository.dart';
 export 'presentation/state/anonymous_subscriptions_cubit.dart';
 export 'presentation/widgets/community_drawer.dart';
 export 'presentation/widgets/community_header/community_header.dart';
